+++ scraper/test-new-sources.ts
@@ -20,6 +20,7 @@ import { scrapePrevue } from './engines/prevue';
 import { scrapeSelkirk } from './engines/selkirk';
 import { scrapeHiringPlatform } from './engines/hiringplatform';
 import { scrapeADP } from './engines/adp';
+import { scrapeApplyToEducation } from './engines/applytoeducation';
 import { scrapeBrassRing, scrapeEdmontonPhenom, scrapeHaltonHills, scrapeNipissing, scrapeNorthernCollege, scrapePickering, scrapeStClairCollege, scrapeStLawrenceCollege, scrapeVaughanPL } from './engines/custom';
 
 const REQUIRED_SUCCESSFUL_RUNS = 3;
@@ -116,6 +117,8 @@ const SOURCES = {
     scrapeHiringPlatform(db, context, 'https://orillia.hiringplatform.ca/list/careers', 'City of Orillia'),
   'Credit Valley Conservation': (db: Client, context: BrowserContext) =>
     scrapeADP(db, context, 'https://workforcenow.adp.com/mascsr/default/mdf/recruitment/recruitment.html?cid=628c6eb9-02b8-492d-b651-c4f1a4220344&ccId=9201548179279_3&lang=en_CA&selectedMenuKey=CurrentOpenings', 'Credit Valley Conservation'),
+  'Toronto Catholic District School Board': (db: Client, context: BrowserContext) =>
+    scrapeApplyToEducation(db, context, 'https://www.tcdsb.org/page/jobs', 'Toronto Catholic District School Board'),
   'Government of Alberta': (db: Client, context: BrowserContext) =>
     scrapeJobs2Web(db, context, 'https://jobpostings.alberta.ca/go/All-Jobs-GoA/2617217/', 'Government of Alberta'),
   'Ontario Energy Board': (db: Client, context: BrowserContext) =>